feat: remove PoC code and move implementation to `EuiTableStickyScrollbar` component and the usage to `EuiTable` for best compatibility
